+++ CHANGELOG.md
@@ -24,7 +24,8 @@ New value objects:
 * Pixels
 
 New methods:
-* Frame.scroll()
+* quixote.browser.canScroll()
+* Frame.scroll() (Compatibility note: fails fast on Mobile Safari)
 * Frame.getRawScrollPosition() (Compatibility notes: IE 8 shim, Mobile Safari scroll container)
 
 Rename: